修改了模板

+++ htadmin/src/main/java/com/lz/ht/dao/DeptMapper.java
@@ -14,7 +14,7 @@ public interface DeptMapper {
 
     void updateById(Dept record);
 
-    int deleteById(Integer DeptId);
+    int deleteById(Long DeptId);
 
     List<Dept> findAll();
