+++ htadmin/src/main/java/com/lz/ht/service/RoleService.java
@@ -11,7 +11,7 @@ public interface RoleService {
 
     void updateById(Role record);
 
-    int deleteById(Integer RoleId);
+    int deleteById(Long RoleId);
 
     List<Role> findAll();
