+++ htadmin/src/main/java/com/lz/ht/service/UserService.java
@@ -11,7 +11,7 @@ public interface UserService {
 
     void updateById(User record);
 
-    int deleteById(Integer UserId);
+    int deleteById(Long UserId);
 
     List<User> findAll();
