+++ htadmin/src/main/java/com/lz/ht/service/serviceImpl/SysConfigServiceImpl.java
@@ -30,7 +30,7 @@ public class SysConfigServiceImpl  implements  SysConfigService {
     }
 
     @Override
-    public int deleteById(Integer SysConfigId) {
+    public int deleteById(Long SysConfigId) {
         return  sysConfigMapper.deleteById(SysConfigId);
     }
